backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-apps==0.1.0
django-cors-headers==3.6.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
